Virtual Private Cloud (VPC) - Explained
The presentation explains the VPC in simple terms making it easy to understand and easy to follow.

With Amazon Virtual Private Cloud (Amazon VPC), you can launch AWS resources in a logically isolated virtual network you've defined. This virtual network resembles a traditional network that you'd operate in your own data center, with the benefits of using the scalable infrastructure of AWS. The diagram shows an example of VPC. The VPC has one subnet in each Availability Zone in the Region, EC2 instances in each subnet, and an internet gateway to allow communication between the resources in your VPC and the internet. For more information, see Amazon Virtual Private Cloud (Amazon VPC).